ASACORE-723 Changed the debug flag option for building on windows
Typically Windows builds use the /Zi flag when doing a debug build. This places
all of the debug information into a separate file with the pdb extension. When
building with scons using the --jobs option multiple processes will build the
code at once.  Then both processes will attempt to write into the same pdb file
this results in an error. By changing the debug flag to /Z7 it will instead
place the debug information into the obj files instead of a separate pdb file
this limits some of the things that are capable in a debugger tool however few
developers know the difference.

Change-Id: I0e01b81bbd61633bb062d06f2157c7b7c894ac88
Signed-off-by: George Nash <georgen@qce.qualcomm.com>