Added AJ_NVRAM_Peek() API and other nvram changes
The peek API returns a pointer to NVRAM without copying it. This is needed by AllJoyn.js but is also generally useful.
Changed AJ_NVRAM_GetSizeRemaining() to do memory compaction if needed. Without this the function is not particularly useful for determining if a requested NVRAM allocation is possible.
Tidied up some comments.

Change-Id: Ic2ac6eed9a80819932e6d3b8cf5e614f14e45bc6
Signed-off-by: gburns <gburns@qce.qualcomm.com>