ASACORE-896 Thin client secure sample client doesn't wait for authentication to complete
The SecureClientECDHE sample has a bug which causes it to begin sending pings before the authentication has completed. This causes errors in the console output until the authentication is completed, at which point the ping will be sent successfully.

I also added a help message if valid parameters aren't provided, and an error message if sending the ping is unsuccessful.

Change-Id: Id2ecf74088c8b98b664285d8afd436af7f3a57f9
Signed-off-by: Rob Smith <robs@microsoft.com>