AJTCL line endings cause WS checker to report incorrectly
Sometimes on windows the whitespace checker will report
that every single file has whitespace errors. WS=detail
shows an empty line when ran. This is caused by the line
endings being incorrect for the system being compiled on
(windows). Changing ajuncrustify "newlines" to auto fixes
the problem.
Tested and confirmed that it works on windows and linux

Change-Id: I43cfebc4bf31fa3439207d246978f5647a90abe2
Signed-off-by: James Prestwood <jprestwo@qce.qualcomm.com>