A ctypes-based Python wrapper for the thin library
This implements a simplified programming environment that presents
AllJoyn object interfaces as Python objects. Method calls, signals,
and properties are set up as decorated Python method calls. Interfaces
are defined as a list of Python objects rather than XML.

Support for simple bus attachments and a limited message loop are
included. To avoid deadlock, handlers for incoming messages should not
make method calls to remote objects. The marshaling and unmarshaling
capabilities are also limited and may not work with large messages.

Change-Id: I93306c974aa070841de85759289bf2b23b15e2b4
Signed-off-by: Mathew Martineau <mathewm@qce.qualcomm.com>