Add support for logging debug output to a file
Provides target-specific support for Linux and OpenWRT platforms for logging debug output to a file.
This is primarily to support running AllJoyn.js as a daemon on these platforms. This is only
applicable for platforms that have file systems so there is no intent to support this capability on
embedded MCU platforms.

Change-Id: I895ad09121587d6b4518b2ebc3db70ab3cca85cf
Signed-off-by: Greg Burns <gburns@qce.qualcomm.com>