ASACORE-953 - Add About Client with Announce processing and handler registration
The AboutClient solution consists of three (3) layers that hide the 
underlying core infrastructure and facilitate application development:

1) Helper functions for unmarshaling the raw Announce signal
2) Helper functions that encapsulate Announce relevance checking
similar to the ASL AnnounceRegistrar that is now being absorbed into 
the BusAttachment.
3) Helper functions for applications that wish to connect as a client 
to a peer and need to make a decision on whether the peer is relevant 
according published interfaces and metadata (AnnouncedData). This is the 
use case for tracking a by its AppId, DeviceName, AppName, Manufacturer, 
etc...).

Change-Id: I6c7ce2d6ab3be48bfbb54f638eb371356583d8ee
Signed-off-by: Eyal Ilsar <eyali@qce.qualcomm.com>