ASACORE-737: print out the wall clock time in the log statement.
When outputting debug statements, use the AJ_GetDebugTime() function to retrieve the time to use. If the function does not return AJ_OK then the elapsed
time is used

For the linux platform, if ER_DEBUG_EPOCH=1 then AJ_GetDebugTime() will return the seconds since the epoch. For all other platforms, AJ_GetDebugTime()
is defined to AJ_ERR_RESOURCES so that elapsed time will be used.

Change-Id: I303f9d4d022f3eb7c50a9c24cf4654b6022913dd
Signed-off-by: Mike Zou <lzou@qce.qualcomm.com>