ASACORE-2068 Full conversation hash
Introduce auth version 4 which adds the following to the conversation
hash to protect them:
     * Header bytes for each phase of the conversation
       (ExchangeGuids, ExchangeSuites, etc)
     * Proposed auth protocol versions from each side
     * Session key generation parameters
     * Auth suite exchange

Conversation hashes are now independent between each version,
to improve what is hashed in parts of the conversation currently
hashed in V1-3, moving away from transformed inputs currently used and
closer to the bytes on the wire. In V4, the entire message body is
hashed.

Additional changes done "along the way":
     * On Windows, AJ_ASSERT macro is changed to use _ASSERT instead
       of assert; on Windows, assert just aborts the process, whereas
       _ASSERT allows trapping into the debugger directly at the point
       of the failed assertion, making it more useful during
       development.
     * Method of computing header pad length moved into a macro rather
       than be duplicated multiple places in the code.
     * Miscellaneous swallowed/ignored error codes now checked.

Change-Id: Ic43b6d3d6ddd0d54b1a2ea6b04363de08987bf90
Signed-off-by: Kevin Kane <kkane@microsoft.com>