ASACORE-2193: Free SHA context when auth context is discarded
The authentication process maintains a hash which can be discarded
depending on negotiated parameters or unsuccessful connections. It
used to be sufficient to zero out the SHA context, but in order to
properly clean up resources it is now necessary to make sure
AJ_SHA256_Init() calls are matched with AJ_SHA256_Final().

The "keepAlive" parameter has been removed from AJ_SHA256_GetDigest()
since calling the function with keepAlive=0 was nearly identical to
AJ_SHA256_Final(), and to make it easier to avoid resource leaks by
properly matching calls to Init and Final.

Context structures are now zeroed out when they are no longer in use.

Change-Id: I1dfe1792b54b7027286ce9b3aab3742a1786fc50
Signed-off-by: Mathew Martineau <mathewm@qce.qualcomm.com>