ASACORE-2231 Use EKUs to indicate certificate type
Update certificate code in various places to use EKUs to indicate if
a certificate is an identity, membership, or unrestricted
certificate instead of a custom extension. Change certificate
ingestion and validation methods to insist leaf certificates are
not unrestricted.

Add AJ_X509FreeDecodedCertificateChain to free the returned chain from
AJ_X509DecodeCertificateChainPEM, instead of duplicating the
implementation in multiple places.

Add a DecodeAndVerifyCertificateChainTest to exercise only the
certificate parsing code.

Copy the client certificates and private key over from
AuthListenerECDHETest in the standard client since they follow the
correct EKU usage.

Fix a bug in AJ_X509DecodeCertificatePEM where it didn't then
decode the underlying DER and populate the rest of the structure.

Change-Id: Iab940ea667439bc532ffab4597633d807424d12a
Signed-off-by: Kevin Kane <kkane@microsoft.com>