ASASCORE-2272: Add APIs to support async method replies
This patch adds a new API to capture the reply context and close the method call message.
The reply context can be used at a later time to compose a reply or error response to
the method call. A call using this new mechanism was added to the basic_service sample.

Change-Id: I3cc220e9b3dc2b6d91f3aa18529b9ba236ae257c
Signed-off-by: gburns <gburns@qce.qualcomm.com>