ASACORE-2233 Include error messages in conversation hash
Follow the principle of: if it appears on the wire during an auth
conversation, hash it. This includes fatal and non-fatal error
messages.

Add HASH_MSG_UNMARSHALED and HASH_MSG_MARSHALED constants for use
as the last parameter to AJ_ConversationHash_Update_Message instead
of having 0 or 1, for clarity in the code.

Change-Id: I8871a65c4cf0b49137651fbfe9d4a1b6a2498487
Signed-off-by: Kevin Kane <kkane@microsoft.com>