ASACORE-2287 Do not require EKUs in end entity certificates
To enable compatibility with pre-15.09 peers, remove the requirement
that end entity certificates have a specific EKU. Transitive EKU
requirements remain, but as pre-15.09 certificate chains were issued
with no EKUs, this will not be a concern.

Modify the SecureClientECDHE sample to support presenting a previous
certificate chain that was not issued with EKUs.

Modify the SecureClientECDHE and SecureServiceECDHE samples to
explicitly call the chain validation code to exercise it.

Change-Id: I7aa628f4332f588982ff7a896bf5f1fef5581f39
Signed-off-by: Kevin Kane <kkane@microsoft.com>