Clarify that no parameters are allowed for AJ_SHA256_Init
AJ_SHA256_Init was previously modified to return a pointer rather than
take one as a parameter. However, the doxygen documentation was not
updated and the use of empty parens in the function declaration would
still allow the old parameter to be passed in (even if it wasn't
used). Now the documentation is fixed and no parameters are allowed.

Change-Id: If435be97ef9607f70e5c8dec6e8625e7bafd5151
Signed-off-by: Mathew Martineau <mathewm@qce.qualcomm.com>