ASACORE-2389: Send an unencrypted error reply if key not found
If a method call was received and there is no key available for it,
the method call could not be decrypted and there is also no way to
authenticate the error response. Send an unencrypted error reply
(rather than nothing) in this case, while making sure other errors
are still authenticated.

Change-Id: Icf01a445894678610ae4c2139d9846ac311a6e20
Signed-off-by: Mathew Martineau <mathewm@qce.qualcomm.com>