ASACORE-2406 Preserve p2p sessioncast behavior under Security 2.0
This patch adds session bookkeeping to AJ_BusAttachment. This is
generally useful for applications, but is used in particular to preserve
the functional equivalence of sessioncast signals in point to point
sessions and unicast signals over the same session.

With Security 2.0 enabled, those signals (which have the exact same
effect if security is disabled) are treated differently: for unicast
signals, policy is enforced whereas for sessioncast signals a different
(or no policy at all) is enforced. We restore equivalence here by
transparently rewriting sessioncast signals over p2p sessions as unicast
signals.

Change-Id: Idfd520ef20d01d501b041d49d828b83e61b3c79d
Signed-off-by: Meggan Ehret <meggan@qeo.org>