ASACORE-2350 Fix PSK authentication when using V1 app callback
Don't refer to a stack-allocated array holding the PSK after it has
gone out of scope when an app uses the AJ_BusSetPasswordCallback
method of providing the PSK and then engages in a V4 auth
conversation.

So as not to rely on the app keeping buffers alive, make local copies
of the PSK hint and key data.

Zero local buffers containing the PSK when no longer needed.

Compute the PSK verifier using the old method when in an auth
conversation with version < V4.

Change-Id: I77db8366b41310bed91f40b8ec5ca1fe04e481ec
Signed-off-by: Kevin Kane <kkane@microsoft.com>