ASACORE-2532: Assert in aj_net.c causing failures
The txData buffer size was increased to 3Kb, which
causes as assert in the ARDP code to fire.  The fix
for ASACORE-2446 was to remove the assert, which
was done for src/target/Linux/aj_net.c, but not for
the corresponding Windows file. This change removes
it from src/win32/aj_net.c.  The same assertion
is not present for other platforms.

Change-Id: Id70abe202d067d0aad6dfb5f4279724bd63b6000
Signed-off-by: gregz <gregz@microsoft.com>
(cherry picked from commit 6548b41376245b5cc7dc21ebb80cf9a34c7efe91)