ASACORE-2612: Do not call the auth listener callback for an invalid cert
Restore v15.04 auth listener behavior where the auth listener callback
is only invoked for valid certificate chains. If the chain validates,
but can't be verified against a policy, the auth listener callback is
invoked so the application can perform further checks on the
certificate chain.

Chain validation is dependent upon the authentication version in use
for the connection being established and is not available at the
application layer, so the application does not know all of the
parameters to pass to AJ_X509VerifyChain().

Change-Id: I630810d1c4dc00ccab6d24c1efa55f54287fbe02
Signed-off-by: Mathew Martineau <mathewm@qce.qualcomm.com>