ASACORE-2599 Notify the app when changing settings
Add StartManagement/EndManagement methods to the security management
interface, that Security Manager can call to notify the target app
about changes.

Also add application callbacks for StartManagement/EndManagement.
A StartManagement callback means that the Security Manager is about
to start updating the security configuration of the current
application. The app should gracefully shut down its sessions
when such a transition occurs. The app can re-establish its sessions
safely after it gets an EndManagement callback.

StartManagement is a not a persisted state, so re-starting the app
moves it back into "management not started" state.

Change-Id: I9c7602fa052eb9314d3de0880208feb5c8f482b6
Signed-off-by: Dan Mihai <Daniel.Mihai@microsoft.com>