ASACORE-2177: Fixing introspection on root path "/"
Introspection in Thin Core requires special handling when the service
path of the app is set to root ("/"). Specifically, this affects
nameChange_client in AJSCL when used with AJTCL's version of
signal_service--the introspection result doesn't contain any
interfaces. Due to the peculiarity of the ChildPath() function, any
path appears as a child of "/", which results in an improper call to
ExpandInterfaces(). Added code to handle "/" as a special case.

Change-Id: I351c2e661ae4b35c3550e9b693b93c4b58575a27
Signed-off-by: Andrey Krokhin <akrokhin@affinegy.com>