ASACORE-2177: Fixing introspection on root path "/"
Introspection in Thin Core requires special handling when the service
path of the app is set to root ("/"). Specifically, this affects
nameChange_client in AJSCL when used with AJTCL's version of
signal_service--the introspection result doesn't contain any
interfaces. Due to the peculiarity of the ChildPath() function, any
path appears as a child of "/", which results in an improper call to
ExpandInterfaces(). Added code to handle "/" as a special case.

Change-Id: I6efdbafc42bb5ba5bd2be77128743cf72a81b6b2
Signed-off-by: Andrey Krokhin <akrokhin@affinegy.com>