ASACORE-2917 Better mutli-homed support for Linux
This commit selects the interface that hosts the default route as the
default DNS interface.
It also re-writes the mDNS packet to use the IP address of the
interface that a broadcast is sent on.
Without this, the route receives queries which it cannot respond to,
and when a valid query is received, it may ignore it as being
part of the same burst.

Change-Id: Iee707077736ace913ddeb6e1cf4114312d2eade8
Signed-off-by: brad <brad@beechwoods.com>